+++ lib/option/some.test.ts
@@ -3,11 +3,11 @@ import { Some } from "./some.ts";
 import { Ok } from "../result/mod.ts";
 
 Deno.test("#isSome", () => {
-  assertEquals(new Some(1).isSome, true);
+  assertEquals(new Some(1).isSome(), true);
 });
 
 Deno.test("#isNone", () => {
-  assertEquals(new Some(1).isNone, false);
+  assertEquals(new Some(1).isNone(), false);
 });
 
 Deno.test("#unwrap", () => {